// src/App/Core/application/dto/establishment/post-establishment.dto.ts
export interface IPostEstablishmentsForm {
    fullname:              string;
    phone_number:          number;
    email:                 string;
    establishment_name:    string;
    establishment_address: string;
    password:              string;
}


export interface IPostEstablishments {
    fullname:              string;
    phone_number:          number;
    email:                 string;
    role_id:               number;
    establishment_name:    string;
    establishment_address: string;
    password:              string;
}